[AVR] Rename the 'none_base' target spec module to 'freestanding_base'
