+++ src/librustc_target/spec/mod.rs
@@ -54,6 +54,7 @@ mod arm_base;
 mod cloudabi_base;
 mod dragonfly_base;
 mod freebsd_base;
+mod freestanding_base;
 mod fuchsia_base;
 mod haiku_base;
 mod hermit_base;
@@ -65,7 +66,6 @@ mod linux_kernel_base;
 mod linux_musl_base;
 mod msvc_base;
 mod netbsd_base;
-mod none_base;
 mod openbsd_base;
 mod redox_base;
 mod riscv_base;
